refactor(dataset): restructure services into orchestration layer
